Pull in LLVM r203311 and r211435.
r211435
Legalizer: Add support for splitting insert_subvectors.
We handle this by spilling the whole thing to the stack and doing the
insertion as a store.

PR19492. This happens in real code because the vectorizer creates
v2i128 when AVX is enabled.

r203311
ISel: Make VSELECT selection terminate in cases where the condition
type has to be split and the result type widened.

When the condition of a vselect has to be split it makes no sense
widening the vselect and thereby widening the condition. We end up in an
endless loop of widening (vselect result type) and splitting (condition
mask type) doing this.  Instead, split both the condition and the
vselect and widen the result.